requirements.txt
aiogram
aiogram_dialog
APScheduler
asyncpg
environs
nats-py
openai
ormsgpack
greenlet
pydantic
SQLAlchemy
pytz

